pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "chromadb>=1.5.0",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.8",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
[tool.setuptools.packages.find]
where = ["."]
